+++ packages/article-skeleton/__tests__/android/article-images.android.test.js
@@ -1,0 +1,3 @@
+import shared from "../images.native";
+
+shared();