+++ packages/article-skeleton/__tests__/ios/article-header-with-style.ios.test.js
@@ -1,0 +1,3 @@
+import shared from "../header-with-style.native";
+
+shared();