+++ packages/ssr/src/lib/ads/make-author-profile-ad-config.js
@@ -1,0 +1,20 @@
+const { defaultClient, defaultServer } = require("./base-ad-config");
+const basePageTargeting = require("./base-page-targeting-config");
+
+module.exports = {
+  defaultClient: () => ({
+    ...defaultClient,
+    pageTargeting: {
+      ...basePageTargeting,
+      cont: "listing"
+    },
+    slotTargeting: {
+      path: "profile/lucy-fisher",
+      section: "profile/lucy-fisher",
+      slot: "home",
+      zone: "profile"
+    }
+  }),
+
+  defaultServer: () => ({ ...defaultServer })
+};