+++ android-app/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.10.22"></a>
+## [0.10.22](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.10.21...@times-components/android-app@0.10.22) (2018-11-22)
+
+**Note:** Version bump only for package @times-components/android-app
+
+
+
+
+
+<a name="0.10.21"></a>
+## [0.10.21](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.10.20...@times-components/android-app@0.10.21) (2018-11-21)
+
+**Note:** Version bump only for package @times-components/android-app
+
+
+
+
+
+<a name="0.10.20"></a>
+## [0.10.20](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.10.19...@times-components/android-app@0.10.20) (2018-11-21)
+
+**Note:** Version bump only for package @times-components/android-app
+
+
+
+
+
 <a name="0.10.19"></a>
 ## [0.10.19](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.10.18...@times-components/android-app@0.10.19) (2018-11-20)