+++ packages/article-main-standard/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.2.3"></a>
+## [0.2.3](https://github.com/newsuk/times-components/compare/@times-components/article-main-standard@0.2.2...@times-components/article-main-standard@0.2.3) (2018-11-22)
+
+**Note:** Version bump only for package @times-components/article-main-standard
+
+
+
+
+
+<a name="0.2.2"></a>
+## [0.2.2](https://github.com/newsuk/times-components/compare/@times-components/article-main-standard@0.2.1...@times-components/article-main-standard@0.2.2) (2018-11-21)
+
+**Note:** Version bump only for package @times-components/article-main-standard
+
+
+
+
+
+<a name="0.2.1"></a>
+## [0.2.1](https://github.com/newsuk/times-components/compare/@times-components/article-main-standard@0.2.0...@times-components/article-main-standard@0.2.1) (2018-11-21)
+
+**Note:** Version bump only for package @times-components/article-main-standard
+
+
+
+
+
 <a name="0.2.0"></a>
 # [0.2.0](https://github.com/newsuk/times-components/compare/@times-components/article-main-standard@0.1.5...@times-components/article-main-standard@0.2.0) (2018-11-20)