+++ packages/error-view/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.1.50"></a>
+## [2.1.50](https://github.com/newsuk/times-components/compare/@times-components/error-view@2.1.49...@times-components/error-view@2.1.50) (2018-11-21)
+
+**Note:** Version bump only for package @times-components/error-view
+
+
+
+
+
 <a name="2.1.49"></a>
 ## [2.1.49](https://github.com/newsuk/times-components/compare/@times-components/error-view@2.1.48...@times-components/error-view@2.1.49) (2018-11-20)