+++ packages/markup/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="3.2.21"></a>
+## [3.2.21](https://github.com/newsuk/times-components/compare/@times-components/markup@3.2.20...@times-components/markup@3.2.21) (2018-11-21)
+
+**Note:** Version bump only for package @times-components/markup
+
+
+
+
+
 <a name="3.2.20"></a>
 ## [3.2.20](https://github.com/newsuk/times-components/compare/@times-components/markup@3.2.19...@times-components/markup@3.2.20) (2018-11-20)