+++ packages/tealium/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.6.12"></a>
+## [1.6.12](https://github.com/newsuk/times-components/compare/@times-components/tealium@1.6.11...@times-components/tealium@1.6.12) (2018-11-21)
+
+**Note:** Version bump only for package @times-components/tealium
+
+
+
+
+
 <a name="1.6.11"></a>
 ## [1.6.11](https://github.com/newsuk/times-components/compare/@times-components/tealium@1.6.10...@times-components/tealium@1.6.11) (2018-11-19)